Fixed subprocess execution problem when the command line is given as a string, not list of arguments (subprocess documentation is somewhat ambiguous about that).
Added support for multiple invocations of pre- and postprocessor.

Modified processing of value sets; each of the sets gets a unique id
from the user (not an option name), which can be placed arbitrarily
in the application command line, as well as pre- and postprocessors.
This eliminates the need for old option "-k" and is much more flexible.

Unified representation of exclusions, so that any non-string sequences 
are possible to match.



git-svn-id: https://svn.cct.lsu.edu/repos/projects/parallex/trunk/hpx@1163 5079c8e4-9419-0410-93bb-9e5ee278c886