Updates to help HPX run on Mac OS X.
CMake scripts were updated so that they try to match from a list of supported boost versions. Essentially, if you have anything from boost 1.33 to 1.39 it will find it. And, when 1.40 come out, we only have to update "set(Boost_ADDITIONAL_VERSIONS ...)" line in CMakeLists.txt.

I also added a check so that we do not build the 'amr' example on Macs, since that was breaking the build process.



git-svn-id: https://svn.cct.lsu.edu/repos/projects/parallex/trunk/hpx@1570 5079c8e4-9419-0410-93bb-9e5ee278c886