Adds a sketch of the ParalleX Data Model. This is meant to describe the type of information we want to collect from an application running in HPX.
The basic idea is that threads can have zero or more children, but only one parent; actions belong to only one component type, and components can have zero or more actions associated with them during a run; and, localities, threads, and actions are associated by events that happen during the simulation. This is sufficient for capturing the causal relationship between actions in the application.



git-svn-id: https://svn.cct.lsu.edu/repos/projects/parallex/trunk/hpx@1775 5079c8e4-9419-0410-93bb-9e5ee278c886