Adds the critcally acclaimed "Please Just Plot" (pjplot) utility by Chris Michael.
Straight from the sources mouth: """Please Just Plot.  Module for generating plots hassle-free.

This module provides operations to help plot data easily.  It is best when used
as a quick and dirty way to plot, as it becomes more tedious to use the more
particular one is about how a plot looks.  Focus is on the data and ease of 
use.

The only plotting program in the implementation so far is gnuplot.  Though
efforts to support many terminal types may be added, use of terminal postscript
is encouraged.  The module is currently in early development, so a lot has not
been implemented.  Look for the "IMP" tag within the documentation to note what
is on the backburner.  If you notice something is missing and would like to
notify me, e-mail me at cmichael@cct.lsu.edu with the name "pjplot" on the
subject line.

IMP: Document gnuplot parameters.

Some examples of usage:

 Simple:
  #Prints the given points to a file named "Test Plot.ps"
  d1 = Data2D("Test Data One", "X", "Y")
  d2 = Data2D("Test Data Two", "X", "Y")

  d1.point(0, 1)
  d1.point(1, 2)
  d1.point(2, 3)
  d2.point(0, 1)
  d2.point(1, 1)
  d2.point(2, 2)

  p = Gnuplot("Test Plot")
  p.data(d1, d2)
  p.plot()
  p.write()

  pWithLines = Gnuplot("Test Plot with Lines")
  pWithLines.dataStyle(d1, withs="lines")
  pWithLines.dataStyle(d2, withs="lines")
  pWithLines.plot()
"""



git-svn-id: https://svn.cct.lsu.edu/repos/projects/parallex/trunk/hpx@2477 5079c8e4-9419-0410-93bb-9e5ee278c886