Adds "thunk" LCO and a new version of the Fibonacci example that uses thunks instead of eager_futures.
A thunk is a construct for delayed evaluation. The interface is identical to that of the eager_future. The difference is that the action is not invoked when the future is created. Instead, the action is invoked when the get() method is used to retrieve the result of the action. Subsequent calls to get() return the value directly.

The thunk-ified version of the fibonacci2 simply replaces
    typedef hpx::lcos::eager_future<...>
with
    typedef hpx::lcos::thunk<...>
and the rest stays the same. This is not an exciting example of thunks, as they cause a depth-first traversal of the "fibonacci action tree". One thing that is interesting is that we see a slight speedup versus the eager_future version, when run with "-t 1"; this is probably caused by the fact that the runtime is never managing more tasks than the one it is currently executing. Another interesting tidbit is that the running time increases with larger values of "-t", even though there is never more than one action running at any time, and that that action never schedules more than one subsequent action.



git-svn-id: https://svn.cct.lsu.edu/repos/projects/parallex/trunk/hpx@2517 5079c8e4-9419-0410-93bb-9e5ee278c886