Moves HPX runtime setup and options processing to separate "hpx_init.hpp" header; updates fibonacci2 to use new hpx_init facility.
The purpose of this is to reduce the burden on the applications programmer by encapsulating general runtime setup that will be the same across all (or most) application. This includes runtime options handling and helper functions such as "split_ip_address". This also has the added benefit of standardizing the set of runtime options (e.g., setting number of OS threads and selecting scheduling policy) across all application codes.



git-svn-id: https://svn.cct.lsu.edu/repos/projects/parallex/trunk/hpx@2530 5079c8e4-9419-0410-93bb-9e5ee278c886