Adds a new Thunk LCO with test.
This is something of a first-class "anytime" future (Phillip's term). You can create a thunk with an associated action and arguments. The action can be invoked by an call to "trigger" the thunk. A call to "get" the value of the thunk will invoke the action if the action was not already invoked, and it will suspend the calling thread pending the availability of the value. The value can be "set"; this action will be resume all threads waiting on the value of the thunk.

The thunk behaves like an eager future if you immediately trigger the action after creating it. It behaves like a lazy future if it is never explicitly triggered (i.e., the call to "get" the value triggers the action). But you are free to play with the eager or laziness of the thunk by varying when you trigger it (that's why it's an anytime future).

The thunk is first-class-ish because you can get the GID of the thunk, pass that to other actions, return it from other actions, and invoke any of the thunk method (trigger, get, and set) from within any context (e.g., a thread on the same locality or a thread on a different locality).



git-svn-id: https://svn.cct.lsu.edu/repos/projects/parallex/trunk/hpx@2577 5079c8e4-9419-0410-93bb-9e5ee278c886