Adds a get_option() helper method to hpx_init.
The get_option() method encapsulates getting a program option from the commandline and from a configuration file. A call to

    get_option(vm, "foo", x);

will set 'x' to the value of the 'foo' commandline option, if one was provided. A call to 

    get_option(vm, "foo", x, "application.my_fancy_app.bar");

will set 'x' to the value of the 'foo' commandline option, if one was provided, or to the value associated with the 'bar' key in some configuration file defining 'my_fancy_app's run parameters.



git-svn-id: https://svn.cct.lsu.edu/repos/projects/parallex/trunk/hpx@2606 5079c8e4-9419-0410-93bb-9e5ee278c886