Adds a multilocality test; a utility class for generating random numbers while multithreading; and, adds an hpx_finalize() method in hpx_init.hpp.
The multilocality test is the classic "calculate Pi by throwing darts at the unit circle" example. You can set the number of throws and the granularity for a run. Granularity is interpreted as the number of throwers per locality. So, if you are using 8 localities and throw 1024 darts with granularity 2, you'll get 2 throwers per locality, each throwing 128 darts.



git-svn-id: https://svn.cct.lsu.edu/repos/projects/parallex/trunk/hpx@2622 5079c8e4-9419-0410-93bb-9e5ee278c886