Adds --clean option to hpx_run.py and updates multilocality tests to use it.
The --clean <name> option kills any stray HPX instances running before it starts a cluster run. It is smart enough to only kill HPX instances on the nodes that it is about to spawn runtime instances to, but too dumb to look for anything but the name provided to the option.

Using this brings the multilocality test up from 20% passing to 70% passing.



git-svn-id: https://svn.cct.lsu.edu/repos/projects/parallex/trunk/hpx@2750 5079c8e4-9419-0410-93bb-9e5ee278c886