Comments out logging of direct actions.
Logging direct actions had the unfortunate result of logging all "console_logging_actions". A small application, spawning less than 10 PX threads, was capable of generating over 2.5 million console_logging_action messages it had to be killed for lack of doing any useful work.

I left the code commented out so that it is easier to restore if a workaround is found.



git-svn-id: https://svn.cct.lsu.edu/repos/projects/parallex/trunk/hpx@2753 5079c8e4-9419-0410-93bb-9e5ee278c886