Adds support for adding examples on a per build basis, using the HPX_WITH_EXAMPLES CMake option.
"To use this, just add -DHPX_WITH_EXAMPLES="<ex>[;<ex>]*" to your favorite cmake command. For example, if you wanted to build HAD and the continuation-passing examples, you would add 

    -DHPX_WITH_EXAMPLES="had_amr;cps"

The benefit is that for out-of-source builds, you do not have to touch the source to build out different examples. If there were multiple components, the same strategy could be applied there.



git-svn-id: https://svn.cct.lsu.edu/repos/projects/parallex/trunk/hpx@2787 5079c8e4-9419-0410-93bb-9e5ee278c886