Adds support on Linux systems for specifying a simple socket/core system topology and selecting a method of mapping OS threads to core.
To use this feature, you need to write a ".hpx.ini" file in the directory where you start the application. The file should have at least the following section:

    [system_topology]
    num_blocks = 2
    num_cores = 4
    thread_map = block_striped

The num_blocks option defaults to 1 and is meant to roughly denote the number of sockets in the system. The num_cores option defaults to the total number of cores detected, and is meant to denote the number of cores per socket. The options for thread_map are "linear" and "block_striped". The linear map simply maps an OS thread with id TID to core number (TID % total_num_cores). This was, and still is, the default behavior. The block_striped map assigns OS threads to cores in a striped fashion. The following illustrates the two different mappings for a run with 4 OS threads on a system with 2 blocks and 4 cores per block.

  blocks/cores      [ 0 1 2 3 ] [ 4 5 6 7]
  linear map          0 1 2 3     - - - -
  block_striped map   0 2 - -     1 3 - -



git-svn-id: https://svn.cct.lsu.edu/repos/projects/parallex/trunk/hpx@2791 5079c8e4-9419-0410-93bb-9e5ee278c886