Fixes build for case where one does not set "HPX_USE_ITT".
I don't know what the possible range of values are for HPX_USE_ITT, but only checking for "!= 0" was not sufficient to keep the build from breaking when not setting HPX_USE_ITT at all. If you update this, make sure it does not break the build for people who do not know what ITT is.



git-svn-id: https://svn.cct.lsu.edu/repos/projects/parallex/trunk/hpx@2810 5079c8e4-9419-0410-93bb-9e5ee278c886