Avoid unnecessary instantiations in is_action
The `is_action` trait required a complete type in its defaulted template argument. This caused unnecessary instantiations when used in async/apply SFINAE checks, even when there is an explicit specialization for a particular action type. Delaying the base template implementation delays instantiations, which are avoided if a specialization is preferred.