Add support for compilers that have thread_local
Clang exposes the ability to test for the C++ feature `thread_local` via
the `__has_feature` macro.

Here we test for the presence of the macro and the feature in two
separate steps to avoid the preprocessor bailing out on invalid syntax.

As preprocessors replace before evaluating, naively going:
`#if defined(__has_feature) && __has_feature(cxx_thread_local)`
will result in the preprocessor seeing `0 && 0(0)` if the compiler
doesn't have Clang-style feature testing.