Add hpx::concurrent::unordered_set a simple wrapper around a std::unordered_set
This has been added as a quick fix to get CERN CMS-SW code compiling with
HPX in place of TBB. It has not been thoroughly tested and should not be
used without checking the behaviour carefully.

A mutex is used to wrap read and write (modifiers) of the set, access
to iterators should be avoided in thread critical places.