Update CMake policies to use CMP0060 NEW
CMake 3.3 introduces CMP0060 which attempts to always use absolute
library paths when linking.

The new behavior resolves a corner case in the old implicit-linking
behavior in which the user may have conflicting versions of a library
(e.g., Boost) installed to both system and custom paths and attempts to
disambiguate with environment variable LIBRARY_PATH.

The old behavior of CMP0060 would implicitly link the libraries with
`-l<lib>` because the path was specified via LIBRARY_PATH and the
compiler could see it. The non-use of an absolute path potentially links
the wrong library version from the system path. The new behavior avoids
this problem by simply using the absolute library path.