Working policy for parcel coalescing. However, it deadlocks because the parcel coalescer registers a callback on the configuration settings to be notified when the values change. Unfortunately, the notification process leads to deadlock when the thread tries to acquire a spinlock that it already owns.
Full description:
The coalescing_message_handler.cpp code sets a callback function for the
“hpx.plugins.coalescing_message_handler.num_messages” value to be
notified when it changes. When we change it from the policy, that callback
function tries to read it - requesting the lock in the process - but the “set”
code isn’t done with it yet.
https://github.com/STEllAR-
GROUP/hpx/blob/master/plugins/parcel/coalescing/coalescing_message_handler
.cpp#L143

Here is where the callback function is called during the “set” program
flow:
https://github.com/STEllAR-GROUP/hpx/blob/master/src/util/ini.cpp#L484

So... one suggestion would be that for the callback function - rather than
get the new value right then, a flag should be set to get a fresh value
the next time the value is needed. The buffer should also be flushed
and reset to the new values.