Redesign of parcelport to avoid maps
Use Heller's approach of having a sender and receiver responsible for
each side of a transfer and use the pointer as the context in fi_xxx
calls so that the object can e retrieved and used directly.
Saves using maps to look up objects.

Some reworking of the header to minimize the amount of data we send
when we do not actually need to.