Reduce header size, simplify rma of message region, add counters
The messages size is reduced significantly to only send those parts
that are neccesary for a particular message type (with/without zero
copy regions etc).

The message region is added to the rma regions list and retrieved
by the receiver without extra message region code other than what is
neccessary for encoding/decoding of the parcel.

Counters are added for messages sent/received, rma operations
and deletes, that are useful for debugging and performance monitoring

Use boost:small_vector to avoid some allocations for rma region storage