Add TLS bitmap for get_area_membind_nodeset and numa domain queries
The hwloc api requires a bitmap to store the numa domain in
and this should not be allocated/freed every time which is an
expensive operation. This adds a thread local bitmap so that
queries to can reuse one each time. Making it thread local
enables different tasks to query memory binding simultaneously.