Support gcc/clang. Accept command-line options. Add daint clang cmake settings
Allow either gcc/clang compiler for machine daint, edit scripts to
change paths to build dirs depending on compiler to allow one or both
to be used at the same time.

Currently, a random compiler is picked for each PR test.

Allow debug/slurm/machine/etc to be passed in on command line - still
checks $env{var} but can be overriden.

Lots of fixes to cmake setup to ensure that options with spaces are
correctly passed to ctest and then to cmake (CMAKE_CXX_FLAGS was not being
correctly handled).