Add new overload of async post to handle passing of schedule hints
The executor can hold priority/stacksize information, but the scheduler
hints are generated on the fly, so cannot be contained in the executor.
To support the passage of hints, we add an overload to the post function
that detects the presence of a hint and passes it correctly through to
the scheduler.

Add override to some functions to avoid accidental redefinitions.