Fixing potential stack overflow for dataflow
When hpx::launch::sync is passed as the execution policy for long dataflow
chains with almost always ready futures, we might run into a stack overflow due
to recursion. This patch fixes it by detecting if there is enough stack space
left and creating a new thread if there isn't.