Fix libfabric PP for N>1 localities on the same node
Using the TCP parcelport with two localities on the same node caused
problems because the libfabric address vector lookup was using only
the ip address. This fix adds the port number to the map key so that
multiple localities on the same node can be supported.

Moving to an index based lookup of locality->fabric address would
be the next obvious step.