Fixing multiple defines of APEX_HAVE_HPX
APEX_HAVE_HPX has to be defined in CMakeLists.hpx of APEX - otherwise, APEX doesn't know to include hpx/config.h.  If it is defined in both places, it gets defined twice and generates warnings.  If it isn't defined in CMakeLists.hpx, then hpx/config.h doesn't get included, and APEX isn't configured correctly.  Also changing version of APEX to be develop to get the latest APEX changes which includes making sure this define is defined.  Before HPX is released, APEX should be tagged and that tag should be used instead of develop.