Add print utility for DartMeasurentTime to lightweight_test
Fixup tests that currently print the time to use the new util
and remove obsolete benchmark commandline param to print time.

The tests will now always run the benchmark and print the time.

Note that names should not include '_' underscores as they are not
plotted by CDash (Stencil3Full is ok but Stencil3_Full is not).