Fix a problem with scope of CMAKE_CXX_FLAGS and hpx_add_compile_flag
CMAKE_CXX_FLAGS_XXX is only valid at directory scope, so flags added
in subdirs are not passed up to the parent. Modify the
add_compile_flag macro to set a global property with flags that can
be read back in the base dir and passed to the executables/libs.

This fixes PR #3439

In the future a target_compile_option would be preferable, but
some more fixes would be required.