Add an empty INTERFACE target to pass compile flags to tests
An INTERFACE target can have no source files, and no library
but passes properties on to targets that depend on it.
We add flags such as -Werror=XXX to the hpx_test_flags target
and make all tests depend on it.

The flags are not visible to user code via the build or install dirs
when they link against the main hpx library that has the same flags
set using the PRIVATE option.