Replace most macros with functions
Functions are much safer to use as they introduce a
separate variable scope.

This commit just substitutes macro definitions with the equivalent
function definitions. It does not touch the function bodies except
to restore the correct behavior. In addition, the commit fixes
a scoping issue in `add_hpx_config_test`.

The remaining macros can not be replaced easily without an
extensive rewrite. Since the commit is already rather large,
I will submit these changes as separate commits.

To verify that the behavior of the build system has not changed,
I have compared the generated compiler invocations and the hashes
of every object file with their old values. The modified build
system produces bit-for-bit identical results for all files (except
version.cpp.o) if the default settings are used.