Remove the '_lib' suffix from all library targets
Having to use different target names depending on the context
is confusing and error-prone. Above all, it makes refactoring the
build system a pain.

Since the suffix is not necessary to disambiguate the existing targets,
this commit simply drops it to unify the target names.

A follow-up commit will do the same for all executable targets.