Link against the C++-enabled MPI libraries
As MPI's C header file (unfortunately) also contains the C++ interface,
only linking against the C libraries might result in unresolved symbols.
Since we can not reliably disable the C++ interface, we have to include
the C++ parts of MPI even if they are not used.

This fixes #3475.

The commit also removes the broken fix.