Adding support for gitlab runners
This adds support for gitlab runners and enables a (semi modular)
setup to add new configurations. This will allow for a basis to add
more PR based tests for more configurations.

The whole workflow was slightly changed:
    - The tests are split up between unit, regression, performance and
    examples. Due to the enhanced infrastructure, we don't need to split
    them anymore
    - Checking the code with clang-tidy is enabled by default. A
    .clang-tidy file has been added for the clang-tidy configuration
    - We build both, a debug and release version. Only the debug version
    is currently pushed to dockerhub on the master branch

This is to replace the slow circle-ci setup (which has been replaced
with a dummy config)