Towards Modularized HPX
This commit adds the basic infrastructure to develop modular HPX
libraries. That is, that each library in the libs subdirectory is
supposed to be a self contained CMake project, with clear dependencies.

The created libraries will get automatically included in the
documentation, that includes API documentation and separate section.
All headers will get automatically tested to be self-contained.