Refactor the generation of the package config files
Previously, HPX used the global variables and directory properties
to construct the package config files (CMake, pkgconf and Bazel),
which would leads to inconsistencies if the requirements was only
added to a specific target.

The new version of the code computes the settings
for the non-CMake config files by walking the dependency graph
starting from a suitable target, e.g. hpx::application, to
construct the transitive usage requirements for a specific setup.
Special CMake constructs like generator expressions are handle
according to their semantics if necessary. Since CMake now solely
relies on the interface properties of the targets, the generated
config files are always in sync with the build system.

Note: The current version of this commit is highly experimental and
requires a major clean-up. Currently, it does not generate the config
files for Bazel.