Use atomic in cancelable_action_client example to avoid exceptions in exception handler
The example sometimes terminates because of an exception thrown while handling a
previous exception. Specifically the interrupted thread throws an exception when
calling suspend. The reset_id struct destructor is then called, which locks a
spinlock which, if it has to wait long enough, calls suspend and then calls
interruption_point. At this point the thread has been interrupted and another
exception will be thrown. This commit uses an atomic id to avoid calling suspend
while the thread is exiting/handling the exception.