Merge the executor-enabled overloads of shared_future<>::then
The separate overload for thread executors caused problems since
the implementation of `then` requires a `scheduled_executor` and not
just an `executor`.

The merged overload still does not test if the user-provided executor
does actually satisfy the requirements of a `scheduled_executor`.
Since future<>::then does exhibit the same issue, this should be
fixed in a separate commit.

This commit also adds a regression test demonstrating the issue.

This fixes issue #3634.