Fixing yet another race in future_data
Run the completion handlers as static functions.
With this change, completion handlers can not have
exceptions that escape the handler as this would lead
to silently swallow the exceptions.
Previously, this would have triggered a "promise_already_satisfied"
error. This is wrong behavior, exceptions escaping completion
handling should be considered a bug. always.