Retry migration operation if a global race is detected.
The (global) race happens in situations where migration operations are triggered
concurrently with actions to be executed on the same object. In this case it can
happen, that actions waiting in AGAS for a ongoing migration operation are released
once this migration has finished. Now, if a new migration operation is started
before the released actions are able to pin the object, then the actions will
pin the object regardless of the new migration operation leading to an exception
being thrown (pin_count > 1).

This situation can't be resolved other then to handle the exception and to retry
the migration. This change set implements this.