EXPAND CONCEPT CHECKING FOR FORWARD ITERATORS
This commit adds a specialization to the `satisfy_traversal_concept` for
the `boost::forward_traversal_tag` provided by the `boost::iterator`
library.

Forward traversal cannot be robustly verified via concept emulation due
to the ambiguities between iterfaces specified by the `ForwardIterator`
and `InputIterator` concepts in the C++20 ranges TS. That said, any
class that satisfies the `BidirectionalIterator` concept is certain to
also fullfill the `ForwardIterator` concept.

The specialization established in this commit leverages that fact to
provide partial concept checking for `ForwardIterator`.