Remove unnecessary barriers in thread test and fix rare deadlock
Barriers were there to ensure that join does not throw. However, join is only an
interruption point for the parent thread, not the thread being interrupted, so
the barriers are not needed.

In addition this fixes a bug which would be caused by the `b->wait` in
`interruption_point_thread` also acting as an interruption point (the barrier
suspends on a condition variable). This is rare but possible. When this happens
the barrier `wait` function throws an exception (`thread_interrupted`) but
cannot decrement the total count that it was waiting for. This eventually leads
to the barrier destructor waiting indefinitely for all threads to enter the
barrier.