Fixing #3792
This fixes the problems reported in #3792. The main issue lies within an
implementation detail of the barrier. The rank of a barrier is
completely decoupled from the locality id. For that reason, the barrier
root is always rank 0. If no barrier with rank 0 is given, we get the
reported hang. The added test shows both a workaround to that issue and
adds an additional barrier constructor that takes a vector of ranks, and
the rank of the given barrier instantiation.