Fix git_external fetching of tags
Tags can't be updated to the latest version with `git rebase FETCH_HEAD` so
don't do it if we're on a tag (or plain commit; anything but a branch).

Also clean up some of the tag handling for APEX and hpxMP (just use the main
HPX_WITH_X_TAG option directly as the tag, instead of defining a separate
variable).