Remove dependency on threadmanager from hpx::thread
hpx::thread constructor used to schedule a thread on the current pool if
in an hpx::thread, and on the default pool otherwise. This creates an
unnecessary dependency on the threadmanager. This commit requires that
the constructor is called on an hpx::thread and schedules it on the
current thread pool. It also adds an additional constructor which takes
an explicit thread pool as an argument to cover the case of constructing
an hpx::thread outside of an hpx::thread.