Fix initialization of global_num_tasks in parallel_executor
Initializing the static variable global_num_tasks with
hpx::get_os_thread_count() could lead to deadlocks when accessing the
configuration option which contains the number of OS threads is contended. In
this case the thread can be suspended while holding a lock for the static
variable. This makes the variable non-static and gets the number of threads
from the thread pool if inside an HPX thread. Otherwise it gets the number of
threads for the default pool.