Rename HPX_LIBRARIES to prepare the export of the variable
We want to add back all the imported target libraries inside an
`HPX_LIBRARIES` variable to give time to external projects to pass
to target directives. We were still using this var in HPX to link
to base libraries like rt or dl. So we now renamed it
`HPX_BASE_LIBRARIES` in prevision of exporting the `HPX_LIBRARIES`
var with all the imported interface libraries.