Add scheduling hints to hello_world_distributed
This both makes sure tasks are scheduled sooner on the correct worker
threads and avoids certain infinite loops in which the foreman thread
always gets scheduled on an odd-numbered thread and the last worker task
always gets scheduled on an even-numbered thread, or vice versa. With an
even number of worker threads this can lead to the worker thread never
being scheduled on the correct thread because of round-robin scheduling.
This commit adds scheduling hints to the worker tasks to ensure this
will not happen.