Only delete tasks on the thread that created them
This commit cleans up the management/lifetime of tasks by ensuring
tasks are only reused when they remain on the same thread so
thread heaps are not protected by a lock any more.
Stolen tasks are simply deleted when finished with and not
added to queues for reuse.

Some queue indexing and thread numbering/tracking issues have been
cleaned up in the process.