Fix launch process test with APEX enabled
The test segfaulted trying to access parent task information from
another locality. This happened because parent task handling was enabled
when the initial number of localities was 1. However, even when the
initial number of localities is 1 additional localities can connect
later (as in the launch process test) leading to APEX trying to acccess
task information from another locality. This changes the behaviour to
only enable parent task handling when networking is disabled.