Add the callback to init the rp's thread_pools
The final goal is to handle the resource partitioner (rp) creation
in the run_or_start method, we'll just pass the resource
partitioner init parameters to the hpx::init() function so that it
then takes care of initializing the rp. We use a callback in case
we want to specify the thread_pools creation or else.