Fix a race condition and clean-up usage of scheduler mode
The pool and the scheduler both kept copies of the scheduler
mode and it was set on startup in different places. Some user
code that wished to set the scheduler mode to a non default
value at the start of hpx_main suffered from a race that allowed
the pool to overwrite the mode flags.

This commit removes the copy of the scheduler mode that was held
by the pool as well as some unnecessary set/get calls to the mode.

This also tidies up a lot of calls to if (get_scheduler_mode() & xxx)
and replaces them with a has_scheduler_mode(xxx) call that makes the
code a little more readable and maintainable.