Fix performance regression in block_executor
In the move from a thread_executor to thread_pool_executor as the underlying
executor for block_executor, the spawning mechanism for bulk task changed from
eagerly creating tasks (run_now = true) to only creating task descriptions
(run_now = false). This changes the default priority of tasks created using the
block_executor to high, which restores the behaviour of immediately creating
threads. It also customizes the hierarchical spawning of tasks from the
thread_pool_executor to allow giving a threshold for when to do hierarchical
spawning. The default threshold leads to non-hierarchical spawning in most
cases.