Add new cudaEvent_t based polling for cuda futures
The existing cuda futures use callbacks to set the future
but cuda also supports the creation+recording of events
followed by an event Query to see if the async operation has
completed.

This patch adds support for creation of futures using event
and polling within the hpx scheduling loop - it closely follows
the mmodel used by the asynchronous mpi support in hpx
but the polling loop is slightly simper since cudaEventQuery
only supports polling a single event at a time
(whereas MPI_Testany allows a vector of requests to be tested).