Allow combined MPI and/or CUDA polling for futures
The polling functions will hardly ever change during normal
operation, so to avoid taking a lock, we use an function pointer
for each of the polling functions for MPI and CUDA separately.
This means they can be set/cleared by any thread and when unchanged
no cache lines are affected and overheads are minimized.