Allow limiting executor to use external counting mechanism
The limiting executor does not work with some special executors
like the MPI one since it returns immediately but leave a task
that is 'detached' from the runtime (the counters count up and down
almost immediately, but the task completes later).

To solve this we add a specialization inside the limiting executor
counting mechanism that allows the executor being wrapped to return
a counter that can be used to throttle tasks.

Any executor that implements the "in_flight_estimate" member can
override the counting mechanism to control throttling.

Currently we only handle a single counter so multiple mpi executors
would reference the same internal mpi counter and it would not be
possible to throttle different mpi executors spearately.

Simplify the mpi_ring_async_test to use a limiting executor at
the top level and thereby reduce resource consumption by only
keeping N mpi requests in flight at once.

Note that the limiting executor is only used for the outer iteration
of sending, otherwise a deadlock would occur if a receive could
not trigger a send.