Make parallel_executor use hierarchical spawning
Alias thread_pool_executor to parallel_executor since the latter now
replaces the functionality of the former. Constructors are backwards
compatible for the commonly used options. The old num_spread and
num_tasks options are no longer arguments to the parallel_executor
constructors and are replaced by the hierarchical_threshold option. The
thread_pool_executor constructor taking a scheduler_base is also no
longer available as it was only used internally.