Add workaround for nvcc compilation
The for_loop_compute test fails to compile without the workarounds in this
commit. The error message is "error: cannot use an entity undefined in device
code" without any mention of what entity is undefined. This workaround is as
specific as I could get without knowing what undefined entity is causing the
error.