Update dataflow to allow customizing dataflow finalization stage
Customization is necessary to allow out-of-band signalling of future
completion. In the normal case the dataflow implementation handles setting the
future to ready, but for e.g. CUDA and cuBLAS executors the completion happens
later, and a custom mechanism is required to be able to set the future to ready
once an asynchronous operation has completed.