Wait for thread counts to indicate idleness multiple times before shutting down
Add helper yield_while variants which require the predicate to return
false multiple times before returning from the loop. This helps reduce
the false positive rate on the raciness of comparing thread counts to
the background thread count.

This also adds functions to check the number of work items left to be
processed by the CUDA and MPI polling functionality. The counts returned
from these are also non-deterministic (unless protected with significant
locks) and so the previous measures help against this as well.

The number of times the yield_while predicate has to return false
consecutively is configurable with the hpx.shutdown_check_count
configuration variable.