Fixing CXX17_NODISCARD_ATTRIBUTE
The cmake tests for CXX17_NODISCARD_ATTRIBUTE are not actually performed but the required cxx17_nodiscard_attribute.cpp is present.
Suprisingly they are missed out in HPX_PerformCxxFeatureTests.cmake so I have just added it.
This solves warning `warn_unused_result` applies only to function types, when compiling with c++17 and later.