Add execution::keep_future to pass the future instead of its value to continuations when futures are used as senders
keep_future passes the future instead of the value to help manage future
lifetimes. Especially shared_future | on | ... is not possible without
this helper if the wrapped value in the future is noncopyable (since
shared_future::get returns a const& and on cannot store a const&, nor
does it need to keep the operation state alive long enough).