Fix sliding semaphore test
The semaphore has to be kept alive long enough for all tasks to get to
signal it. The task that releases the waiting thread may not be the last
task to signal the semaphore, and if that happens the semaphore will go
out of scope too early, unless we wait for all tasks to finish first.