Fix annotate_function
std::strings as annotations were not correctly handled in the
constructor. std::strings would use the constructor expecting functions,
which attempts to get annotations with get_function_annotation, which
returns "<unknown>" if there is no specialization. This adds an explicit
constructor for std::strings and disables the constructor taking
functions for std::strings.