Fix block_os_threads_1036 test
The started variable may be accessed by the "blocker" threads after it
has gone out of scope, leaving the blocker thread reading garbage from
started and the loop

while (started->load() != 1) continue;

never returning and the test hanging.

This adds a barrier at the end of the test to ensure that no variables
in the "blocker" threads are accessed after they go out of scope from
hpx_main. The rest of the test is meant to use never-suspending
primitives, but at the end of the test we can use a barrier which may
suspend.