Update future_await_traits
Remove allocator template parameters from the promise types, and instead
only use hpx::util::internal_allocator. Since we know the allocator we
can construct it for allocation/deallocation instead of storing it.
operator new for the promise_type only allocates now, and does not
attempt to construct the promise_type.