Define HPX_COMPUTE_CODE in builds with SYCL
This resolves a compilation issue where timestamp_cuda is required for a
certain build configuration (using ONEAPI/SYCL/dpcpp with an enabled
CUDA backend and specific GPU architecture) but not found. The
underlying issue is that the include for the timestamp_cuda header is
protected with a ifdef HPX_COMPUTE_CODE and the header was thus not
included  - hence the timestamp method was not found and
the code did not compile in this configuration!

This commit resolves the issue! Not by changing or removing the ifdef
timestamp guard, but instead by defining HPX_COMPUTE_CODE for SYCL
builds in the compiler_specific header. This ensures that the behaviour
for non-SYCL builds stays the same as before!